RANDOMWALK-20: separate users from members
