+++ src/main/java/ru/random/walk/club_service/controller/UserController.java
@@ -3,14 +3,10 @@ package ru.random.walk.club_service.controller;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.graphql.data.method.annotation.Argument;
-import org.springframework.graphql.data.method.annotation.MutationMapping;
 import org.springframework.graphql.data.method.annotation.QueryMapping;
 import org.springframework.stereotype.Controller;
-import ru.random.walk.club_service.mapper.MemberMapper;
 import ru.random.walk.club_service.model.entity.AnswerEntity;
 import ru.random.walk.club_service.model.entity.ClubEntity;
-import ru.random.walk.club_service.model.entity.MemberEntity;
-import ru.random.walk.club_service.model.graphql.types.MemberRole;
 import ru.random.walk.club_service.model.graphql.types.PaginationInput;
 import ru.random.walk.club_service.util.StubDataUtil;
 
@@ -22,8 +18,6 @@ import java.util.UUID;
 @Slf4j
 @AllArgsConstructor
 public class UserController {
-    private final MemberMapper memberMapper;
-
     @QueryMapping
     public List<ClubEntity> getUserClubs(
             @Argument UUID userId,
@@ -55,23 +49,5 @@ public class UserController {
                 StubDataUtil.formAnswerEntity(),
                 StubDataUtil.membersConfirmAnswerEntity()
         );
-    }
-
-    @MutationMapping
-    MemberEntity changeMemberRole(
-            @Argument UUID clubId,
-            @Argument UUID memberId,
-            @Argument MemberRole role
-    ) {
-        log.info("""
-                        Change member role
-                        for club id [{}]
-                        member id [{}]
-                        with role [{}]
-                        """,
-                clubId, memberId, role
-        );
-        var memberRole = memberMapper.toDomainMemberRole(role);
-        return StubDataUtil.memberEntityWith(memberId, memberRole);
     }
 }
